# src/python/dart/web/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /engine:
    get:
      operationId: listEngines
      summary: "list engines"
      tags:
        - "Engine"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the engines matching the filters or all engines if no filters are specified."
          schema:
            $ref: '#/definitions/PagedEnginesResponse'
    post:
      operationId: createEngine
      summary: "create an engine"
      tags:
        - "Engine"
      parameters:
        - name: "engine"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Engine"
      responses:
        "200":
          description: "the engine"
          schema:
            $ref: '#/definitions/EngineResponse'
  /engine/{engine_id}:
    get:
      operationId: getEngine
      summary: "get an engine instance"
      tags:
        - "Engine"
      parameters:
        - name: "engine_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "the engine"
          schema:
            $ref: '#/definitions/EngineResponse'
    put:
      operationId: updateEngine
      summary: "update an engine"
      tags:
        - "Engine"
      parameters:
        - name: "engine_id"
          type: "string"
          in: "path"
          required: true
        - name: "engine"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Engine"
      responses:
        "200":
          description: "the engine"
          schema:
            $ref: '#/definitions/EngineResponse'
    delete:
      operationId: deleteEngine
      summary: "delete an engine"
      tags:
        - "Engine"
      parameters:
        - name: "engine_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
  /engine/{engine_id}/subgraph_definition:
    get:
      operationId: getEngineSubgraphDefinition
      summary: "return the subgraph definition for the engine"
      tags:
        - "Engine"
        - "SubgraphDefinition"
      parameters:
        - name: "engine_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            $ref: "#/definitions/SubgraphDefinitionResponse"
    post:
      operationId: createEngineSubgraphDefinition
      summary: "create the subgraph definition for the engine"
      tags:
        - "Engine"
        - "SubgraphDefinition"
      parameters:
        - name: "engine_id"
          type: "string"
          in: "path"
          required: true
        - name: "subgraph_definition"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/SubgraphDefinition"
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            $ref: "#/definitions/SubgraphDefinitionResponse"
  /engine/action/{action_id}/checkout:
    put:
      operationId: checkoutEngineAction
      summary: "Checkout the action"
      tags:
        - "Engine"
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/ActionContextResponse"
        "400":
          description: "engine action is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "engine or datastore not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /engine/action/{action_id}/checkin:
    put:
      operationId: checkinEngineAction
      summary: "Checkin the action"
      tags:
        - "Engine"
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
        - name: "action_result"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/ActionResult"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
        "400":
          description: "engine action is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "engine or datastore not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /event:
    post:
      operationId: createEvent
      summary: "create an event"
      tags:
        - "Event"
      parameters:
        - name: "event"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Event"
      responses:
        "200":
          description: returns the event
          schema:
            $ref: '#/definitions/EventResponse'
    get:
      operationId: listEvents
      summary: "list events"
      tags:
        - "Event"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the events matching the filters or all events if no filters are specified."
          schema:
            $ref: '#/definitions/PagedEventsResponse'
  /event/{event_id}:
    get:
      operationId: getEvent
      summary: "get an event"
      tags:
        - "Event"
      parameters:
        - name: "event_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: returns the event
          schema:
            $ref: '#/definitions/EventResponse'
    put:
      operationId: updateEvent
      summary: "update an event"
      tags:
        - "Event"
      parameters:
        - name: "event_id"
          type: "string"
          in: "path"
          required: true
        - name: "event"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Event"
      responses:
        "200":
          description: returns the event
          schema:
            $ref: '#/definitions/EventResponse'
    patch:
      operationId: patchEvent
      summary: "patch update an event"
      tags:
        - "Event"
      parameters:
        - name: "event_id"
          type: "string"
          in: "path"
          required: true
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the event
          schema:
            $ref: '#/definitions/EventResponse'
    delete:
      operationId: deleteEvent
      summary: "delete an event"
      tags:
        - "Event"
      parameters:
        - name: "event_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
...
  /subgraph_definition/{subgraph_definition_id}:
    get:
      operationId: getSubgraphDefinition
      summary: "get the subgraph definition"
      tags:
        - "SubgraphDefinition"
      parameters:
        - name: "subgraph_definition_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "the subgraph definition"
          schema:
            $ref: '#/definitions/SubgraphDefinitionResponse'
    delete:
      operationId: deleteSubgraphDefinition
      summary: "delete the subgraph definition"
      tags:
        - "SubgraphDefinition"
      parameters:
        - name: "subgraph_definition_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
...
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
...
  ActionContext:
    type: object
  ActionContextResponse:
    properties:
      results:
        $ref: '#/definitions/ActionContext'
    type: object
...
  ActionResult:
    type: object
  ActionType:
    properties:
      description:
        type: string
      name:
        type: string
      params_json_schema:
        type: object
        x-nullable: true
    type: object
...
  Engine:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/EngineData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  EngineData:
    additionalProperties: false
    properties:
      description:
        type: string
        x-nullable: true
      ecs_task_definition:
        type: object
        x-nullable: true
      ecs_task_definition_arn:
        type: string
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      options_json_schema:
        type: object
        x-nullable: true
      supported_action_types:
        items:
          $ref: '#/definitions/ActionType'
        minItems: 1
        type: array
        x-schema-form:
          title: '{{ value.name }}'
          type: tabarray
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
    required:
    - name
    - user_id
    type: object
  EngineResponse:
    properties:
      results:
        $ref: '#/definitions/Engine'
    type: object
  ErrorResponse:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
  Event:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/EventData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  EventData:
    additionalProperties: false
    properties:
      description:
        type: string
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      state:
        default: INACTIVE
        enum:
        - ACTIVE
        - INACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
    required:
    - name
    - user_id
    type: object
  EventResponse:
    properties:
      results:
        $ref: '#/definitions/Event'
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResponse:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...
  PagedEnginesResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Engine'
        type: array
      total:
        type: integer
    type: object
  PagedEventsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Event'
        type: array
      total:
        type: integer
    type: object
...
  SubgraphDefinition:
    type: object
  SubgraphDefinitionResponse:
    properties:
      results:
        $ref: '#/definitions/SubgraphDefinition'
    type: object
...